Migrating more critical tests ahead of release.
This includes QA checking scripts, which have already uncovered one
issue with one of the updated rigs.
