+++ test_files/addon_tests.py
@@ -22,6 +22,11 @@
 # ##### END GPL LICENSE BLOCK #####
 
 
+"""
+DEPRECATED
+"""
+
+
 from contextlib import redirect_stdout
 import filecmp
 import importlib
@@ -48,20 +53,11 @@ class mcprep_testing():
 		self.suppress = True  # hold stdout
 		self.test_status = {}  # {func.__name__: {"check":-1, "res":-1,0,1}}
 		self.test_cases = [
-			self.check_blend_eligible,
-			self.check_blend_eligible_middle,
-			self.check_blend_eligible_real,
-			self.import_world_split,
-			self.import_world_fail,
-			self.import_jmc2obj,
-			self.import_mineways_separated,
-			self.import_mineways_combined,
 			self.name_generalize,
 			self.canonical_name_no_none,
 			self.canonical_test_mappings,
 			self.detect_extra_passes,
 			self.find_missing_images_cycles,
-			self.qa_meshswap_file,
 			self.geonode_effect_spawner,
 			self.particle_area_effect_spawner,
 			self.collection_effect_spawner,
@@ -71,11 +67,6 @@ class mcprep_testing():
 			self.uv_transform_detection,
 			self.uv_transform_no_alert,
 			self.uv_transform_combined_alert,
-			self.test_enable_obj_importer,
-			self.qa_effects,
-			self.qa_rigs,
-			self.convert_mtl_simple,
-			self.convert_mtl_skip,
 		]
 		self.run_only = None  # Name to give to only run this test
 
@@ -436,164 +427,6 @@ class mcprep_testing():
 		# check on image that is packed or not, or packed but no data
 		os.remove(tmp_image)
 
-	def check_blend_eligible(self):
-		from MCprep.spawner import spawn_util
-		fake_base = "MyMob - by Person"
-
-		suffix_new = " pre9.0.0"  # Force active blender instance as older
-		suffix_old = " pre1.0.0"  # Force active blender instance as newer.
-
-		p_none = fake_base + ".blend"
-		p_new = fake_base + suffix_new + ".blend"
-		p_old = fake_base + suffix_old + ".blend"
-		rando = "rando_name" + suffix_old + ".blend"
-
-		# Check where input file is the "non-versioned" one.
-
-		res = spawn_util.check_blend_eligible(p_none, [p_none, rando])
-		if res is not True:
-			return "Should have been true even if rando has suffix"
-
-		res = spawn_util.check_blend_eligible(p_none, [p_none, p_old])
-		if res is not True:
-			return "Should be true as curr blend eligible and checked latest"
-
-		res = spawn_util.check_blend_eligible(p_none, [p_none, p_new])
-		if res is not False:
-			print(p_none, p_new, res)
-			return "Should be false as curr blend not eligible and checked latest"
-
-		# Now check if input is a versioned file.
-
-		res = spawn_util.check_blend_eligible(p_new, [p_none, p_new])
-		if res is not True:
-			return "Should have been true since we are below min blender"
-
-		res = spawn_util.check_blend_eligible(p_old, [p_none, p_old])
-		if res is not False:
-			return "Should have been false since we are above this min blender"
-
-	def check_blend_eligible_middle(self):
-		# Warden-like example, where we have equiv of pre2.80, pre3.0, and
-		# live blender 3.0+ (presuming we want to test a 2.93-like user)
-		from MCprep.spawner import spawn_util
-		fake_base = "WardenExample"
-
-		# Assume the "current" version of blender is like 9.1
-		# To make test not be flakey, actual version of blender can be anything
-		# in range of 2.7.0 upwards to 8.999.
-		suffix_old = " pre2.7.0"  # Force active blender instance as older.
-		suffix_mid = " pre9.0.0"  # Force active blender instance as older.
-		suffix_new = ""  # presume "latest" version
-
-		p_old = fake_base + suffix_old + ".blend"
-		p_mid = fake_base + suffix_mid + ".blend"
-		p_new = fake_base + suffix_new + ".blend"
-
-		# Test in order
-		filelist = [p_old, p_mid, p_new]
-
-		res = spawn_util.check_blend_eligible(p_old, filelist)
-		if res is True:
-			return "Older file should not match (in order)"
-		res = spawn_util.check_blend_eligible(p_mid, filelist)
-		if res is not True:
-			return "Mid file SHOULD match (in order)"
-		res = spawn_util.check_blend_eligible(p_new, filelist)
-		if res is True:
-			return "Newer file should not match (in order)"
-
-		# Test out of order
-		filelist = [p_mid, p_new, p_old]
-
-		res = spawn_util.check_blend_eligible(p_old, filelist)
-		if res is True:
-			return "Older file should not match (out of order)"
-		res = spawn_util.check_blend_eligible(p_mid, filelist)
-		if res is not True:
-			return "Mid file SHOULD match (out of order)"
-		res = spawn_util.check_blend_eligible(p_new, filelist)
-		if res is True:
-			return "Newer file should not match (out of order)"
-
-	def check_blend_eligible_real(self):
-		# This order below matches a user's who was encountering an error
-		# (the actual in-memory python list order)
-		riglist = [
-			"bee - Boxscape.blend",
-			"Blaze - Trainguy.blend",
-			"Cave Spider - Austin Prescott.blend",
-			"creeper - TheDuckCow.blend",
-			"drowned - HissingCreeper-thefunnypie2.blend",
-			"enderman - Trainguy.blend",
-			"Ghast - Trainguy.blend",
-			"guardian - Trainguy.blend",
-			"hostile - boxscape.blend",
-			"illagers - Boxscape.blend",
-			"mobs - Rymdnisse.blend",
-			"nether hostile - Boxscape.blend",
-			"piglin zombified piglin - Boxscape.blend",
-			"PolarBear - PixelFrosty.blend",
-			"ravager - Boxscape.blend",
-			"Shulker - trainguy.blend",
-			"Skeleton - Trainguy.blend",
-			"stray - thefunnypie2.blend",
-			"Warden - DigDanAnimates pre2.80.0.blend",
-			"Warden - DigDanAnimates pre3.0.0.blend",
-			"Warden - DigDanAnimates.blend",
-			"Zombie - Hissing Creeper.blend",
-			"Zombie Villager - Hissing Creeper-thefunnypie2.blend"
-		]
-		target_list = [
-			"Warden - DigDanAnimates pre2.80.0.blend",
-			"Warden - DigDanAnimates pre3.0.0.blend",
-			"Warden - DigDanAnimates.blend",
-		]
-
-		from MCprep.spawner import spawn_util
-		if bpy.app.version < (2, 80):
-			correct = "Warden - DigDanAnimates pre2.80.0.blend"
-		elif bpy.app.version < (3, 0):
-			correct = "Warden - DigDanAnimates pre3.0.0.blend"
-		else:
-			correct = "Warden - DigDanAnimates.blend"
-
-		for rig in target_list:
-			res = spawn_util.check_blend_eligible(rig, riglist)
-			if rig == correct:
-				if res is not True:
-					return "Did not pick {} as correct rig".format(rig)
-			else:
-				if res is True:
-					return "Should have said {} was correct - not {}".format(correct, rig)
-
-	def import_world_split(self):
-		"""Test that imported world has multiple objects"""
-		self._clear_scene()
-
-		pre_objects = len(bpy.data.objects)
-		self._import_jmc2obj_full()
-		post_objects = len(bpy.data.objects)
-		if post_objects + 1 > pre_objects:
-			print("Success, had {} objs, post import {}".format(
-				pre_objects, post_objects))
-			return
-		elif post_objects + 1 == pre_objects:
-			return "Only one new object imported"
-		else:
-			return "Nothing imported"
-
-	def import_world_fail(self):
-		"""Ensure loader fails if an invalid path is loaded"""
-		testdir = os.path.dirname(__file__)
-		obj_path = os.path.join(testdir, "jmc2obj", "xx_jmc2obj_test_1_14_4.obj")
-		try:
-			bpy.ops.mcprep.import_world_split(filepath=obj_path)
-		except Exception as e:
-			print("Failed, as intended: " + str(e))
-			return
-		return "World import should have returned an error"
-
 	def import_materials_util(self, mapping_set):
 		"""Reusable function for testing on different obj setups"""
 		from MCprep.materials.generate import get_mc_canonical_name
@@ -681,30 +514,6 @@ class mcprep_testing():
 
 		# also test that there are not raw image names not in mapping list
 		# but that otherwise could be added to the mapping list as file exists
-
-	def import_jmc2obj(self):
-		"""Checks that material names in output obj match the mapping file"""
-		self._clear_scene()
-		self._import_jmc2obj_full()
-
-		res = self.import_materials_util("block_mapping_jmc")
-		return res
-
-	def import_mineways_separated(self):
-		"""Checks Mineways (single-image) material name mapping to mcprep_data"""
-		self._clear_scene()
-		self._import_mineways_separated()
-
-		res = self.import_materials_util("block_mapping_mineways")
-		return res
-
-	def import_mineways_combined(self):
-		"""Checks Mineways (multi-image) material name mapping to mcprep_data"""
-		self._clear_scene()
-		self._import_mineways_combined()
-
-		res = self.import_materials_util("block_mapping_mineways")
-		return res
 
 	def name_generalize(self):
 		"""Tests the outputs of the generalize function"""
@@ -863,60 +672,6 @@ class mcprep_testing():
 		if res != {}:
 			cleanup()
 			raise Exception("Fake file should not have any return")
-
-	def _qa_helper(self, basepath, allow_packed):
-		"""File used to help QC an open blend file."""
-
-		# bpy.ops.file.make_paths_relative() instead of this, do manually.
-		different_base = []
-		not_relative = []
-		missing = []
-		for img in bpy.data.images:
-			if not img.filepath:
-				continue
-			abspath = os.path.abspath(bpy.path.abspath(img.filepath))
-			if not abspath.startswith(basepath):
-				if allow_packed is True and img.packed_file:
-					pass
-				else:
-					different_base.append(os.path.basename(img.filepath))
-			if img.filepath != bpy.path.relpath(img.filepath):
-				not_relative.append(os.path.basename(img.filepath))
-			if not os.path.isfile(abspath):
-				if allow_packed is True and img.packed_file:
-					pass
-				else:
-					missing.append(img.name)
-
-		if len(different_base) > 50:
-			return "Wrong basepath for image filepath comparison!"
-		if different_base:
-			return "Found {} images with different basepath from file: {}".format(
-				len(different_base), ", ".join(different_base))
-		if not_relative:
-			return "Found {} non relative img files: {}".format(
-				len(not_relative), ", ".join(not_relative))
-		if missing:
-			return "Found {} img with missing source files: {}".format(
-				len(missing), ", ".join(missing))
-
-	def qa_meshswap_file(self):
-		"""Open the meshswap file, assert there are no relative paths"""
-		basepath = os.path.join("MCprep_addon", "MCprep_resources")
-		basepath = os.path.abspath(basepath)  # relative to the dev git folder
-		blendfile = os.path.join(basepath, "mcprep_meshSwap.blend")
-		if not os.path.isfile(blendfile):
-			return blendfile + ": missing tests dir local meshswap file"
-		bpy.ops.wm.open_mainfile(filepath=blendfile)
-		# do NOT save this file!
-
-		resp = self._qa_helper(basepath, allow_packed=False)
-		if resp:
-			return resp
-
-		# detect any non canonical material names?? how to exclude?
-
-		# Affirm that no materials have a principled node, should be basic only
 
 	def geonode_effect_spawner(self):
 		"""Test the geo node variant of effect spawning works."""
@@ -1217,184 +972,6 @@ class mcprep_testing():
 		invalid, invalid_objs = detect_invalid_uvs_from_objs([lava_obj, water_obj])
 		if invalid is False:
 			return "Combined lava/water should still alert"
-
-	def test_enable_obj_importer(self):
-		"""Ensure module name is correct, since error won't be reported."""
-		bpy.ops.preferences.addon_enable(module="io_scene_obj")
-
-	def qa_effects(self):
-		"""Ensures that effects files meet all QA needs"""
-
-		basepath = os.path.join("MCprep_addon", "MCprep_resources", "effects")
-		basepath = os.path.abspath(basepath)  # relative to the dev git folder
-
-		bfiles = []
-		for child in os.listdir(basepath):
-			fullp = os.path.join(basepath, child)
-			if os.path.isfile(fullp) and child.lower().endswith(".blend"):
-				bfiles.append(fullp)
-			elif not os.path.isdir(fullp):
-				continue
-			subblends = [
-				os.path.join(basepath, child, blend)
-				for blend in os.listdir(os.path.join(basepath, child))
-				if os.path.isfile(os.path.join(basepath, child, blend))
-				and blend.lower().endswith(".blend")
-			]
-			bfiles.extend(subblends)
-
-		print("Checking blend files")
-		if not bfiles:
-			return "No files loaded for bfiles"
-
-		issues = []
-		checked = 0
-		for blend in bfiles:
-			print("QC'ing:", blend)
-			if not os.path.isfile(blend):
-				return "Did not exist: " + str(blend)
-			bpy.ops.wm.open_mainfile(filepath=blend)
-
-			resp = self._qa_helper(basepath, allow_packed=True)
-			if resp:
-				issues.append([blend, resp])
-			checked += 1
-
-		if issues:
-			return issues
-
-	def qa_rigs(self):
-		"""Ensures that all rig files meet all QA needs.
-
-		NOTE: This test is actually surpisingly fast given the number of files
-		it needs to check against, but it's possible it can be unstable. Exit
-		early to override if needed.
-		"""
-
-		basepath = os.path.join("MCprep_addon", "MCprep_resources", "rigs")
-		basepath = os.path.abspath(basepath)  # relative to the dev git folder
-
-		bfiles = []
-		for child in os.listdir(basepath):
-			fullp = os.path.join(basepath, child)
-			if os.path.isfile(fullp) and child.lower().endswith(".blend"):
-				bfiles.append(fullp)
-			elif not os.path.isdir(fullp):
-				continue
-			subblends = [
-				os.path.join(basepath, child, blend)
-				for blend in os.listdir(os.path.join(basepath, child))
-				if os.path.isfile(os.path.join(basepath, child, blend))
-				and blend.lower().endswith(".blend")
-			]
-			bfiles.extend(subblends)
-
-		print("Checking blend files")
-		if not bfiles:
-			return "No files loaded for bfiles"
-
-		issues = []
-		checked = 0
-		for blend in bfiles:
-			print("QC'ing:", blend)
-			if not os.path.isfile(blend):
-				return "Did not exist: " + str(blend)
-			bpy.ops.wm.open_mainfile(filepath=blend)
-
-			resp = self._qa_helper(basepath, allow_packed=True)
-			if resp:
-				issues.append([blend, resp])
-			checked += 1
-
-		if issues:
-			return "Checked {} rigs, issues: {}".format(
-				checked, issues)
-
-	def convert_mtl_simple(self):
-		"""Ensures that conversion of the mtl with other color space works."""
-		from MCprep import world_tools
-
-		src = "mtl_simple_original.mtl"
-		end = "mtl_simple_modified.mtl"
-		test_dir = os.path.dirname(__file__)
-		simple_mtl = os.path.join(test_dir, src)
-		modified_mtl = os.path.join(test_dir, end)
-
-		# now save the texturefile somewhere
-		tmp_dir = tempfile.gettempdir()
-		tmp_mtl = os.path.join(tmp_dir, src)
-		shutil.copyfile(simple_mtl, tmp_mtl)  # leave original intact
-
-		if not os.path.isfile(tmp_mtl):
-			return "Failed to create tmp tml at " + tmp_mtl
-
-		# Need to mock:
-		# bpy.context.scene.view_settings.view_transform
-		# to be an invalid kind of attribute, to simulate an ACES or AgX space.
-		# But we can't do that since we're not (yet) using the real unittest
-		# framework, hence we'll just clear the  world_tool's vars.
-		save_init = list(world_tools.BUILTIN_SPACES)
-		world_tools.BUILTIN_SPACES = ["NotRealSpace"]
-		print("TEST: pre", world_tools.BUILTIN_SPACES)
-
-		# Resultant file
-		res = world_tools.convert_mtl(tmp_mtl)
-
-		# Restore the property we unset.
-		world_tools.BUILTIN_SPACES = save_init
-		print("TEST: post", world_tools.BUILTIN_SPACES)
-
-		if res is None:
-			return "Failed to mock color space and thus could not test convert_mtl"
-
-		if res is False:
-			return "Convert mtl failed with false response"
-
-		# Now check that the data is the same.
-		res = filecmp.cmp(tmp_mtl, modified_mtl, shallow=False)
-		if res is not True:
-			# Not removing file, since we likely want to inspect it.
-			return "Generated MTL is different: {} vs {}".format(
-				tmp_mtl, modified_mtl)
-		else:
-			os.remove(tmp_mtl)
-
-	def convert_mtl_skip(self):
-		"""Ensures that we properly skip if a built in space active."""
-		from MCprep import world_tools
-
-		src = "mtl_simple_original.mtl"
-		test_dir = os.path.dirname(__file__)
-		simple_mtl = os.path.join(test_dir, src)
-
-		# now save the texturefile somewhere
-		tmp_dir = tempfile.gettempdir()
-		tmp_mtl = os.path.join(tmp_dir, src)
-		shutil.copyfile(simple_mtl, tmp_mtl)  # leave original intact
-
-		if not os.path.isfile(tmp_mtl):
-			return "Failed to create tmp tml at " + tmp_mtl
-
-		# Need to mock:
-		# bpy.context.scene.view_settings.view_transform
-		# to be an invalid kind of attribute, to simulate an ACES or AgX space.
-		# But we can't do that since we're not (yet) using the real unittest
-		# framework, hence we'll just clear the  world_tool's vars.
-		actual_space = str(bpy.context.scene.view_settings.view_transform)
-		save_init = list(world_tools.BUILTIN_SPACES)
-		world_tools.BUILTIN_SPACES = [actual_space]
-		print("TEST: pre", world_tools.BUILTIN_SPACES)
-
-		# Resultant file
-		res = world_tools.convert_mtl(tmp_mtl)
-
-		# Restore the property we unset.
-		world_tools.BUILTIN_SPACES = save_init
-		print("TEST: post", world_tools.BUILTIN_SPACES)
-
-		if res is not None:
-			os.remove(tmp_mtl)
-			return "Should not have converter MTL for valid space"
 
 
 class OCOL:
