+++ test_files/world_tools_test.py
@@ -16,7 +16,10 @@
 #
 # ##### END GPL LICENSE BLOCK #####
 
+import filecmp
 import os
+import shutil
+import tempfile
 import unittest
 
 import bpy
@@ -50,6 +53,10 @@ class WorldToolsTest(unittest.TestCase):
         self.assertGreater(
             len(bpy.data.materials), 50, "Should have many mats")
 
+    def test_enable_obj_importer(self):
+        """Ensure module name is correct, since error won't be reported."""
+        bpy.ops.preferences.addon_enable(module="io_scene_obj")
+
     def test_world_import_jmc_full(self):
         test_subpath = os.path.join(
             "test_data", "jmc2obj_test_1_15_2.obj")
@@ -67,6 +74,13 @@ class WorldToolsTest(unittest.TestCase):
             "test_data", "mineways_test_combined_1_15_2.obj")
         self._import_world_with_settings(file=test_subpath)
         self.assertEqual(self.addon_prefs.MCprep_exporter_type, "Mineways")
+
+    def test_world_import_fails_expected(self):
+        testdir = os.path.dirname(__file__)
+        obj_path = os.path.join(testdir, "fake_world.obj")
+        with self.assertRaises(RuntimeError):
+            res = bpy.ops.mcprep.import_world_split(filepath=obj_path)
+            self.assertEqual(res, {'CANCELLED'})
 
     def test_add_mc_sky(self):
         subtests = [
@@ -98,6 +112,89 @@ class WorldToolsTest(unittest.TestCase):
                     self.assertFalse(obj, "Static scn should have no timeobj")
                 else:
                     self.assertTrue(obj, "Dynamic scn should have timeobj")
+
+    def test_convert_mtl_simple(self):
+        """Ensures that conversion of the mtl with other color space works."""
+
+        src = "mtl_simple_original.mtl"
+        end = "mtl_simple_modified.mtl"
+        test_dir = os.path.dirname(__file__)
+        simple_mtl = os.path.join(test_dir, "test_data", src)
+        modified_mtl = os.path.join(test_dir, "test_data", end)
+
+        # now save the texturefile somewhere
+        tmp_dir = tempfile.gettempdir()
+        tmp_mtl = os.path.join(tmp_dir, src)
+        shutil.copyfile(simple_mtl, tmp_mtl)  # leave original intact
+
+        self.assertTrue(
+            os.path.isfile(tmp_mtl),
+            f"Failed to create tmp tml at {tmp_mtl}")
+
+        # Need to mock:
+        # bpy.context.scene.view_settings.view_transform
+        # to be an invalid kind of attribute, to simulate an ACES or AgX space.
+        # But we can't do that since we're not (yet) using the real unittest
+        # framework, hence we'll just clear the  world_tool's vars.
+        save_init = list(world_tools.BUILTIN_SPACES)
+        world_tools.BUILTIN_SPACES = ["NotRealSpace"]
+        print("TEST: pre", world_tools.BUILTIN_SPACES)
+
+        # Resultant file
+        res = world_tools.convert_mtl(tmp_mtl)
+
+        # Restore the property we unset.
+        world_tools.BUILTIN_SPACES = save_init
+        print("TEST: post", world_tools.BUILTIN_SPACES)
+
+        self.assertIsNotNone(
+            res,
+            "Failed to mock color space and thus could not test convert_mtl")
+
+        self.assertTrue(res, "Convert mtl failed with false response")
+
+        # Now check that the data is the same.
+        res = filecmp.cmp(tmp_mtl, modified_mtl, shallow=False)
+        self.assertTrue(
+            res, f"Generated MTL is different: {tmp_mtl} vs {modified_mtl}")
+        # Not removing file, since we likely want to inspect it.
+        os.remove(tmp_mtl)
+
+    def test_convert_mtl_skip(self):
+        """Ensures that we properly skip if a built in space active."""
+
+        src = "mtl_simple_original.mtl"
+        test_dir = os.path.dirname(__file__)
+        simple_mtl = os.path.join(test_dir, "test_data", src)
+
+        # now save the texturefile somewhere
+        tmp_dir = tempfile.gettempdir()
+        tmp_mtl = os.path.join(tmp_dir, src)
+        shutil.copyfile(simple_mtl, tmp_mtl)  # leave original intact
+
+        self.assertTrue(
+            os.path.isfile(tmp_mtl), f"Failed to create tmp tml at {tmp_mtl}")
+
+        # Need to mock:
+        # bpy.context.scene.view_settings.view_transform
+        # to be an invalid kind of attribute, to simulate an ACES or AgX space.
+        # But we can't do that since we're not (yet) using the real unittest
+        # framework, hence we'll just clear the  world_tool's vars.
+        actual_space = str(bpy.context.scene.view_settings.view_transform)
+        save_init = list(world_tools.BUILTIN_SPACES)
+        world_tools.BUILTIN_SPACES = [actual_space]
+        # print("TEST: pre", world_tools.BUILTIN_SPACES)
+
+        # Resultant file
+        res = world_tools.convert_mtl(tmp_mtl)
+
+        # Restore the property we unset.
+        world_tools.BUILTIN_SPACES = save_init
+        # print("TEST: post", world_tools.BUILTIN_SPACES)
+
+        if res is not None:
+            os.remove(tmp_mtl)
+        self.assertIsNone(res, "Should not have converter MTL for valid space")
 
 
 if __name__ == '__main__':
